1. Move one js-file; 2. Fix some styles in one scss-file.
