+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import '../styles/App.css';
+import './styles/App.css';
 
 export const App = () => {
   const [messageList, setMessageList] = useState([]);
