+++ src/styles/App.scss
@@ -1,5 +1,4 @@
 $width__app: 80vw;
-$min_width__app: 900px;
 $width__app_form: 0.9 * $width__app;
 $min_width__app_form: 0.9 * $width__app;
 $width_proportion__input: 0.9;
@@ -7,7 +6,6 @@ $width_proportion__btn: 1 - $width_proportion__input;
 
 .app {
   width: $width__app;
-  min-width: $min_width__app;
   margin: 0 auto;
   background-color: #88bbdd;
   display: flex;
@@ -21,11 +19,13 @@ $width_proportion__btn: 1 - $width_proportion__input;
   height: 5vh;
   width: 100%;
   min-height: 25px;
+  display: flex;
 }
 
 .app__form_input {
   height: 100%;
-  width: 90%;
+  width: 20px;
+  flex-grow: 1;
   border-top-left-radius: 5vw;
   border-bottom-left-radius: 5vw;
   border-right: none;
@@ -40,7 +40,7 @@ $width_proportion__btn: 1 - $width_proportion__input;
 
 .app__form_btn {
   height: 100%;
-  width: 10%;
+  width: 90px;
   border-top-right-radius: 5vw;
   border-bottom-right-radius: 5vw;
   border-left: none;
